// src/app/shared/enum/document-type.enum.ts
export enum DataType {
    Boolean = 1,
    Integer = 2,
    String = 3,
    Double = 4,
    Date = 5,
    Time = 6,
    Datetime = 7,
    Phone = 8,
    Email = 9,
    List = 10
}